Fix class options in cli/profiles
At some point, `remove_class_options` worked like the `private` or
`protected` keywords, and would only remove the options from following
command definitions.

No commands in the cli/profiles class had the --profiles option,
though this class option was removed near the bottom of the class
definition.

Perhaps this is broken in this application because of the way
subcommands are set up? It doesn't seem to be due to subclassing from
SubcommandBase.

At any rate, it is now fixed, though I've introduced a redundant (to
the class_option definition) :profiles_option method definition in
SubcommandBase to do so.